[mod] refactor: move Flask proxy fix to searx.flaskfix module
